requirements.txt
pandas-1.0.3
numpy-1.18.4
xlrd-1.2.0
haversine-2.2.0